+++ Sources/ProjectDescription/Environment.swift
@@ -10,6 +10,7 @@ public struct Environment {
         /// - Parameters:
         ///   - default: default String value to be returned
         /// - Returns: String
+        @available(*, deprecated, message: "Environment variables as optionals are deprecated. Remove question mark to fix deprecation warning.")
         public func getString(default defaultString: String) -> String {
             if case let .string(value) = self { return value }
             return defaultString
@@ -19,6 +20,7 @@ public struct Environment {
         /// - Parameters:
         ///   - default: default Boolean value to be returned
         /// - Returns: Bool
+        @available(*, deprecated, message: "Environment variables as optionals are deprecated. Remove question mark to fix deprecation warning.")
         public func getBoolean(default defaultBoolean: Bool) -> Bool {
             if case let .boolean(value) = self { return value }
             return defaultBoolean
@@ -37,6 +39,26 @@ public struct Environment {
         } else {
             return .string(value)
         }
+    }
+}
+
+extension Optional where Wrapped == Environment.Value {
+    /// Retrieve the Environment value as a string or return the specified default string value
+    /// - Parameters:
+    ///   - default: default String value to be returned
+    /// - Returns: String
+    public func getString(default defaultString: String) -> String {
+        if case let .string(value) = self { return value }
+        return defaultString
+    }
+
+    /// Retrieve the Environment value as a boolean or return the specified default boolean value
+    /// - Parameters:
+    ///   - default: default Boolean value to be returned
+    /// - Returns: Bool
+    public func getBoolean(default defaultBoolean: Bool) -> Bool {
+        if case let .boolean(value) = self { return value }
+        return defaultBoolean
     }
 }